moved FeroxResponse to its own file
