+++ src/filters/mod.rs
@@ -1,14 +1,12 @@
 //! contains all of feroxbuster's filters
-mod wildcard;
-mod status_code;
-mod words;
-mod lines;
-mod size;
-mod regex;
-mod similarity;
-mod container;
-#[cfg(test)]
-mod tests;
+use std::any::Any;
+use std::fmt::Debug;
+
+use crate::ferox_response::FeroxResponse;
+use crate::{
+    config::CONFIGURATION,
+    traits::{FeroxFilter, FeroxSerialize},
+};
 
 pub use self::container::FeroxFilters;
 pub use self::lines::LinesFilter;
@@ -19,10 +17,13 @@ pub use self::status_code::StatusCodeFilter;
 pub use self::wildcard::WildcardFilter;
 pub use self::words::WordsFilter;
 
-use crate::{
-    config::CONFIGURATION,
-    traits::{FeroxFilter, FeroxSerialize},
-    FeroxResponse,
-};
-use std::any::Any;
-use std::fmt::Debug;
+mod wildcard;
+mod status_code;
+mod words;
+mod lines;
+mod size;
+mod regex;
+mod similarity;
+mod container;
+#[cfg(test)]
+mod tests;
